pyproject.toml
[project]
name = "felix-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "watchdog>=6.0.0",
    "websockets>=15.0.1",
]
